extend verify otp variables
